FTL Entries Parser is a wrapper on AST Parser and Transformer
